supabase_version/pyproject.toml
[project]
name = "tarsight"
version = "0.1.0"
description = "Tarsight 测试框架"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "allure-pytest>=2.15.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=0.24.0",
    "pytest-html>=4.1.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
    "aiohttp>=3.10.0",
    "httpx>=0.27.0", # 推荐：异步和同步API统一，更易使用
    "rich>=14.1.0",
    "selenium>=4.15.0", # 用于自动化登录处理��杂验证
    "cryptography>=41.0.0", # 用于配置文件加密
    "python-dotenv>=1.2.1",
    "pandas>=2.3.3",
    "supabase>=2.0.0",
]
[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
